token

+++ src/services/productService.js
@@ -1,13 +1,18 @@
 const API_URL = `${import.meta.env.VITE_API_URL}/products`;
 
+const getToken = () => {
+  const token = localStorage.getItem("token");
+  return token ? `Bearer ${token}` : null;
+};
+
 const handleResponse = async (response) => {
   const data = await response.json();
 
   if (!response.ok) {
-    throw new Error (data.message || "Error en la petición");
-  };
+    throw new Error(data.message || "Error en la petición");
+  }
   return data;
-}
+};
 
 export const getProducts = async () => {
   const response = await fetch(API_URL);
@@ -22,12 +27,17 @@ export const getProductById = async (id) => {
 };
 
 export const createProduct = async (productData) => {
+  const token = getToken();
 
-  
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(API_URL, {
     method: "POST",
     headers: {
       "Content-Type": "application/json",
+      Authorization: token,
     },
     body: JSON.stringify(productData),
   });
@@ -36,10 +46,17 @@ export const createProduct = async (productData) => {
 };
 
 export const updateProduct = async (id, productData) => {
+  const token = getToken();
+
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(`${API_URL}/${id}`, {
     method: "PUT",
     headers: {
       "Content-Type": "application/json",
+      Authorization: token,
     },
     body: JSON.stringify(productData),
   });
@@ -48,8 +65,15 @@ export const updateProduct = async (id, productData) => {
 };
 
 export const deleteProduct = async (id) => {
+  const token = getToken();
+
+  if (!token) {
+    throw new Error("No autorizado");
+  }
+
   const response = await fetch(`${API_URL}/${id}`, {
     method: "DELETE",
+    headers: { Authorization: token },
   });
 
   return handleResponse(response);
